renaming butterfree module
